Code review comments fix
