+++ src/hearings/components/index.ts
@@ -1,15 +1,16 @@
-import {CaseFlagsComponent} from './case-flags/case-flags.component';
+import { CaseFlagsComponent } from './case-flags/case-flags.component';
+import { HearingActualSummaryComponent } from './hearing-actual-summary/hearing-actual-summary.component';
 import { HearingActualsCaseTitleComponent } from './hearing-actuals-case-title/hearing-actuals-case-title.component';
-import {HearingAmendWarningMsgComponent} from './hearing-amend-warning-msg/hearing-amend-warning-msg.component';
+import { HearingAmendWarningMsgComponent } from './hearing-amend-warning-msg/hearing-amend-warning-msg.component';
 import { HearingCaseNameComponent } from './hearing-case-name/hearing-case-name.component';
-import {HearingConfirmationComponent} from './hearing-confirmation/hearing-confirmation.component';
-import {HearingJudgeNameComponent} from './hearing-judge-name/hearing-judge-name.component';
-import {HearingJudgeNamesListComponent} from './hearing-judge-names-list/hearing-judge-names-list.component';
-import {HearingPartiesTitleComponent} from './hearing-parties-title/hearing-parties-title.component';
-import {HearingSummaryComponent} from './hearing-summary/hearing-summary.component';
+import { HearingConfirmationComponent } from './hearing-confirmation/hearing-confirmation.component';
+import { HearingJudgeNameComponent } from './hearing-judge-name/hearing-judge-name.component';
+import { HearingJudgeNamesListComponent } from './hearing-judge-names-list/hearing-judge-names-list.component';
+import { HearingPartiesTitleComponent } from './hearing-parties-title/hearing-parties-title.component';
+import { HearingSummaryComponent } from './hearing-summary/hearing-summary.component';
 import { ListingInformationSummaryComponent } from './listing-information-summary/listing-information-summary.component';
-import {MultiLevelSelectorComponent} from './multi-level-selector/multi-level-selector.component';
-import {PartiesUnavailableDatesComponent} from './parties-unavailable-dates/parties-unavailable-dates.component';
+import { MultiLevelSelectorComponent } from './multi-level-selector/multi-level-selector.component';
+import { PartiesUnavailableDatesComponent } from './parties-unavailable-dates/parties-unavailable-dates.component';
 
 export const components: any[] = [
   CaseFlagsComponent,
@@ -23,17 +24,20 @@ export const components: any[] = [
   ListingInformationSummaryComponent,
   HearingConfirmationComponent,
   HearingAmendWarningMsgComponent,
-  HearingActualsCaseTitleComponent
+  HearingActualsCaseTitleComponent,
+  HearingActualSummaryComponent
 ];
 
 export * from '../components/case-flags/case-flags.component';
-export * from '../components/hearing-parties-title/hearing-parties-title.component';
-export * from '../components/parties-unavailable-dates/parties-unavailable-dates.component';
+export * from '../components/hearing-actual-summary/hearing-actual-summary.component';
+export * from '../components/hearing-actuals-case-title/hearing-actuals-case-title.component';
+export * from '../components/hearing-amend-warning-msg/hearing-amend-warning-msg.component';
+export * from '../components/hearing-confirmation/hearing-confirmation.component';
 export * from '../components/hearing-judge-name/hearing-judge-name.component';
-export * from './multi-level-selector/multi-level-selector.component';
 export * from '../components/hearing-judge-names-list/hearing-judge-names-list.component';
+export * from '../components/hearing-parties-title/hearing-parties-title.component';
 export * from '../components/hearing-summary/hearing-summary.component';
 export * from '../components/listing-information-summary/listing-information-summary.component';
-export * from '../components/hearing-confirmation/hearing-confirmation.component';
-export * from '../components/hearing-amend-warning-msg/hearing-amend-warning-msg.component';
-export * from '../components/hearing-actuals-case-title/hearing-actuals-case-title.component';
+export * from '../components/parties-unavailable-dates/parties-unavailable-dates.component';
+export * from './multi-level-selector/multi-level-selector.component';
+
