--- src/hearings/containers/hearing-actuals/hearing-completed-summary/hearing-completed-summary.component.scss
@@ -1,3 +1,0 @@
-.completed-summary{
-  padding-left: 10px;
-}
