+++ src/hearings/containers/hearing-actuals/index.ts
@@ -3,7 +3,6 @@ import { HearingActualsFinalConfirmationComponent } from './hearing-actuals-fina
 import { HearingActualsTimingComponent } from './hearing-actuals-timing/hearing-actuals-timing.component';
 import { HearingActualsViewEditPartiesComponent } from './hearing-actuals-view-edit-parties/hearing-actuals-view-edit-parties.component';
 import { HearingActualsComponent } from './hearing-actuals.component';
-import { HearingCompletedSummaryComponent } from './hearing-completed-summary/hearing-completed-summary.component';
 import { HearingStageResultComponent } from './hearing-stage-result/hearing-stage-result.component';
 
 export const containers: any[] = [
@@ -13,5 +12,4 @@ export const containers: any[] = [
   HearingStageResultComponent,
   HearingActualsFinalConfirmationComponent,
   HearingActualsViewEditPartiesComponent,
-  HearingCompletedSummaryComponent,
 ];
