+++ src/hearings/containers/view-hearing/hearing-completed-summary/hearing-completed-summary.component.html
@@ -1,0 +1,1 @@
+<exui-hearing-actual-summary></exui-hearing-actual-summary>
