+++ src/hearings/containers/view-hearing/hearing-completed-summary/hearing-completed-summary.component.ts
@@ -1,0 +1,8 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'exui-hearing-completed-summary',
+  templateUrl: './hearing-completed-summary.component.html',
+})
+export class HearingCompletedSummaryComponent {
+}
