+++ src/hearings/containers/view-hearing/index.ts
@@ -1,9 +1,11 @@
 import {HearingCancellationSummaryComponent} from './hearing-cancellation-summary/hearing-cancellation-summary.component';
 import {HearingCancelledSummaryComponent} from './hearing-cancelled-summary/hearing-cancelled-summary.component';
+import { HearingCompletedSummaryComponent } from './hearing-completed-summary/hearing-completed-summary.component';
 import {ViewHearingComponent} from './view-hearing.component';
 
 export const containers: any[] = [
   ViewHearingComponent,
   HearingCancellationSummaryComponent,
-  HearingCancelledSummaryComponent
+  HearingCancelledSummaryComponent,
+  HearingCompletedSummaryComponent
 ];
